Move ObservableProperty, ExtensionsProperty, and Observable parsing code into observables.py to prevent circular imports and fix #23.
