+++ stix2/sdo.py
@@ -4,9 +4,10 @@ import stix2
 
 from .base import _STIXBase
 from .common import COMMON_PROPERTIES
+from .observables import ObservableProperty
 from .other import KillChainPhase
 from .properties import (IDProperty, IntegerProperty, ListProperty,
-                         ObservableProperty, ReferenceProperty, StringProperty,
+                         ReferenceProperty, StringProperty,
                          TimestampProperty, TypeProperty)
 from .utils import NOW
